requirements.txt
django~=4.0
djangorestframework~=3.12
pillow~=9.1
psycopg2-binary~=2.9
reportlab~=3.6